+++ packages/dart_edge_http_server/CHANGELOG.md
@@ -1,3 +1,8 @@
+## 0.3.4
+
+- Update `dart_edge_http_server_runtime` constraint for prebuilt native asset
+  support.
+
 ## 0.3.0
 
 - Declare internal Dart Edge dependencies with the internal hosted registry.